Adding tabulation of values which relate to Einstein's theory
Introducting three things - a textarea for storing a Pandas-friendly JSON to contain some tabulated data, a couple of paragraphs detailing what those tabulated data are for, and a new script to make that text area actually store that Pandas-friendly data. Also modifying the motion script because I need to, since the restart code needs to empty those tabulated values and the restart code is inside the motion code.